+++ action_server/docs/CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## Unreleased
+
+- The response from a run now includes an `"X-Action-Server-Run-Id"` header containing the run id.
+    - This makes it possible to query more information from `api/runs/{run_id}` after the run finishes.
+
 ## 0.0.17 - 2024-01-19
 
 - By default the minimum number of processes is now 2.